+++ src/sdk/models/synergia.ts
@@ -63,8 +63,8 @@ export interface HomeWork {
   CreatedBy: JsonObject | ApiRef | null;
   Date: string;
   Id: number;
-  LessonNo: number;
-  Subject: JsonObject | ApiRef | null;
+  LessonNo: string | number | null;
+  Subject?: JsonObject | ApiRef | null;
   TimeFrom: string | null;
   TimeTo: string | null;
 }